pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "local-agent"
version = "0.1.0"
description = "AI-driven browser automation tool using Claude Computer Use"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "playwright>=1.49.0",
    "anthropic>=0.42.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "Pillow>=11.0.0",
    "typer>=0.15.0",
    "websockets>=14.0",
]

[project.scripts]
local-agent = "cli.client:app"

[tool.hatch.build.targets.wheel]
packages = ["src/local_agent", "cli"]
